Rename `Listener` -> `PullHandler` and other fixes
