+++ sensebook/__init__.py
@@ -1,7 +1,5 @@
 """Making sense of Facebooks undocumented API."""
 
-__version__ = "0.1.2"
-
 from ._utils import (
     default_user_agent,
     parse_form,
@@ -11,9 +9,11 @@ from ._utils import (
     time_from_millis,
     random_hex,
 )
-from ._abc import State, ABCRequest
+from ._abc import State, Request
 
 from ._login import LoginError
-from ._listen import ProtocolError, PullRequest, Listener
+from ._pull_handler import ProtocolError, PullRequest, PullHandler
+
+__version__ = "0.1.2"
 
 __all__ = ()
